cloudrun/requirements.txt
google-cloud-storage==2.7.0
gtts==2.3.1
moviepy==1.0.3
numpy==1.23.5
opencv-python==4.7.0.72
Pillow==9.5.0
Flask==2.3.2
gunicorn==21.2.0